Created a view for the Postit Notes.
Postit Notes view contains a form to submit a Postit Note. The form allows a user to enter a name and text to be submitted on the server.

Postit Notes view also shows all Postit Notes in a reverse order.

Written to be usable for other instances, such as showing notes from a specific user.

Added a link to /postit on the Home Page.